corba: added cmake cache variable PLUGINS_CORBA_SEND_ONEWAY_OPERATIONS to enable or disable oneway operations
If PLUGINS_CORBA_SEND_ONEWAY_OPERATIONS is set to ON, operations that have a collect arity of zero
(void return value and no reference arguments) are sent as oneway calls. The fake SendHandle returned to
the caller will always evaluate to SendSuccess, as if the operation would have been completed successfully.

Note that exceptions in the operation implementation are not forwarded to the caller and print a log message
at the server's side only.

Signed-off-by: Johannes Meyer <johannes@intermodalics.eu>